+++ urlfix/urlfix.py
@@ -1,9 +1,17 @@
 from collections.abc import Sequence
 import os
-from pathlib import Path
 import re
 import urllib.request
 from urllib.request import Request
+import tempfile
+from urllib.error import URLError, HTTPError
+import warnings
+
+
+def file_format(in_file):
+    format_pattern = r'.+\.(\w+)'
+    matches = re.findall(format_pattern, in_file)
+    return matches[0] if len(matches) > 0 else ''
 
 
 class URLFix(object):
@@ -15,9 +23,7 @@ class URLFix(object):
         self.input_file = input_file
         self.output_file = output_file
         # automatically detect input file format
-        format_pattern = r'.+\.(\w+)'
-        matches = re.findall(format_pattern, self.input_file)
-        self.input_format = matches[0] if len(matches) > 0 else ''
+        self.input_format = file_format(self.input_file)
 
     def replace_urls(self, verbose=False, correct_urls=None, inplace=False):
         """
@@ -27,7 +33,7 @@ class URLFix(object):
         :return  Replaces outdated URL and writes to the specified file. It also returns the number of URLs that have
         changed. The latter is useful for tests.
         """
-        if self.input_format not in ("md", "txt"):
+        if self.input_format not in ("md", "txt", "rmd", "Rmd", "rst"):
             raise NotImplementedError(f"File format {self.input_format} is not yet supported.")
         else:
             pass
@@ -35,107 +41,90 @@ class URLFix(object):
         link_text = "[^]]+"
         # Better markdown link matching  taken from https://stackoverflow.com/a/23395483/10323798
         # http:// or https:// followed by anything but a closing paren
-        actual_link = "http[s]?://[^)]+"
-        combined_regex = f"\[({link_text})]\(\s*({actual_link})\s*\)"
+        actual_link = r"http[s]?://[^)|^\s|?<=\]]+"
+        # Need to find more links if using double bracket Markdown hence define single md []() RegEx.
+        single_md = r"\[([^]]+)\]\((http[s]?://[^\s|^\)]+)\)"
+        combined_regex = fr"\[({link_text})\]\(({actual_link})\)\]\((http[s].*)\)|({single_md})"
         # Match only links in a text file, do not text that follows.
         # Assumes that links will always be followed by a space.
-        final_regex = "http[s]?://[^\s]+" if self.input_format == "txt" else combined_regex
+        final_regex = r"http[s]?://[^\s]+" if self.input_format in ["rst", "txt"] else combined_regex
+
+        if self.output_file is None:
+            if not inplace:
+                raise ValueError("Please provide an output file to write to.")
+            else:
+                # Get directory name from input file path
+                output_file = tempfile.NamedTemporaryFile(dir=os.path.dirname(self.input_file), delete=False,
+                                                          mode="w")
+
+        else:
+            if not all(os.path.exists(x) for x in [self.input_file, self.output_file]):
+                raise FileNotFoundError("input_file and output_file should be valid files.")
+
+            output_file = open(self.output_file, "w")
 
         number_moved = 0
         number_of_urls = 0
 
-        if not self.output_file and not inplace:
-            raise ValueError("Please provide an output file to write to.")
-        output_file = self.output_file
-        if inplace:
-            output_file = "replacements_tmp.txt"
-            with open(output_file, "w"):
-                #Create temp file if temp file is not specified
-                pass
-
-        if not all(os.path.isfile(x) for x in [self.input_file, output_file]):
-            raise FileNotFoundError("input_file and output_file should be valid files.")
-
-        with open(self.input_file, "r") as input_f, open(output_file, "w") as out_f:
+        with open(self.input_file, "r") as input_f, output_file as out_f:
 
             for line in input_f:
                 matched_url = re.findall(final_regex, line)
+                # Drop empty strings
+                if self.input_format in ["md", "rmd", "Rmd"]:
+                    matched_url = [list(str(x) for x in texts_links if x != '') for texts_links in matched_url]
+                    for link_texts in matched_url:
+                        if len(link_texts) > 1:
+                            link_texts = link_texts[1:]
+                            # This is used because for some reason we match links twice if single md []()
+                            # This isn't ideal
+                            # TODO: Use better Regular Expression that matches the target links at once
+                            matched_url = list(filter(lambda x: ("https" or "http") in x, link_texts))
+                if len(matched_url) == 0:
+                    # If no URL found, write this line so it is kept in the output file.
+                    out_f.write(line)
+                    pass
+                else:
+                    for final_link in matched_url:
+                        number_of_urls += 1
+                        if isinstance(correct_urls, Sequence) and final_link in correct_urls:
+                            # skip current url if it's in 'correct_urls'
+                            print(f'{final_link} is already valid.')
+                            continue
 
-                if len(matched_url) != 0:
-                    matched_url = matched_url[0][1] if self.input_format == "md" else matched_url[0]
-                    number_of_urls += 1
-
-                    # make sure 'correct_urls' parameter is a sequence
-
-                    if isinstance(correct_urls, Sequence) and matched_url in correct_urls:
-                        # skip current url if it's in 'correct_urls'
-                        print(f'{matched_url} is already valid.')
-                        continue
-
-                    # This printing step while unnecessary may be useful to make sure things work as expected
-                    if verbose:
-                        print(f"Found {matched_url} in {input_f.name}, now validating.. ")
-                    visited_url = urllib.request.urlopen(Request(matched_url, headers={'User-Agent': 'XYZ/3.0'}))
-                    url_used = visited_url.geturl()
-                    if url_used != matched_url:
-                        number_moved += 1
+                        # This printing step while unnecessary may be useful to make sure things work as expected
                         if verbose:
-                            print(f"{matched_url} replaced with {url_used} in {out_f.name}")
-                    out_f.write(line.replace(matched_url, url_used))
+                            print(f"Found {final_link} in {input_f.name}, now validating.. ")
+                        try:
+                            visited_url = urllib.request.urlopen(
+                                Request(final_link, headers={'User-Agent': 'XYZ/3.0'}))
+                            url_used = visited_url.geturl()
 
-        if inplace:
-            os.remove(self.input_file)
-            if verbose:
-                print(f"{self.input_file} removed successfully")
-            os.rename(output_file, self.input_file)
-            if verbose:
-                print(f"{output_file} renamed with {self.input_file} successfully")
+                        except HTTPError as err:
+                            # Put HTTPError before URLError to avoid issues with inheritance
+                            # This may be useful for 4xxs, 3xxs if we get past the URLError
+                            warnings.warn(f"{final_link} not updated, got HTTP error code: {err.code}.")
+                            pass
+
+                        except URLError as err:
+                            warnings.warn(f"{final_link} not updated. Reason: {err.reason}")
+                            # Must be a way to skip, for now rewrite it in there
+                            pass
+
+                        else:
+                            if url_used != final_link:
+                                number_moved += 1
+                                line = line.replace(final_link, url_used)
+                                if verbose:
+                                    print(f"{final_link} replaced with {url_used} in {out_f.name}")
+
+                    out_f.write(line)
+
         information = "URLs have changed" if number_moved != 1 else "URL has changed"
         print(f"{number_moved} {information} of the {number_of_urls} links found in {self.input_file}")
-        return number_moved
+        if inplace:
+            os.replace(out_f.name, self.input_file)
+            if verbose:
+                print(f"Renamed temporary file {output_file} as {self.input_file}")
 
-
-class DirURLFix(object):
-    def __init__(self, input_dir):
-        """
-        :param input_dir: Path to input_dir.
-        """
-        self.input_dir = input_dir
-        self.use_files = Path(self.input_dir)
-
-    def __replace_by_format(self, _format, **kwargs):
-        """
-        :param _format: Input format of the files. Currently supports md and txt
-        :param inplace: Flag for inplace update operation.
-        :return: File with outdated URLs replaced.
-        """
-        for input_file in self.use_files.glob(f'*.{_format}'):
-            input_file = str(input_file) # convert object to path string
-            if "inplace" in kwargs and kwargs["inplace"]:
-                output_file = "replacements_tmp.txt"
-            else:
-                if '_output' in input_file:
-                    print(f"File is a fix of another file: {input_file}")
-                    continue  # skip output files
-
-                output_file = input_file.replace(f'.{_format}', f'_output.{_format}')
-                if os.path.exists(output_file):
-                    print(f"File already fixed: {input_file}")
-                    continue  # skip file that's already been fixed
-
-                with open(output_file, 'w'):
-                    pass  # create an empty output file
-            return URLFix(input_file, output_file).replace_urls(**kwargs)
-
-    def replace_urls(self, **kwargs):
-        if not os.path.exists(self.input_dir):
-            raise OSError('Path does not exist!')
-        if not os.path.isdir(self.input_dir):
-            raise NotADirectoryError('Input path must be a directory!')
-        number_moved = []
-        for _format in ('md', 'txt'):
-            for file_found in list(self.use_files.glob(f"*.{_format}")):
-                number_moved.append(self.__replace_by_format(_format, **kwargs))
-
-        print('All files have been updated, thank you for using urlfix.')
         return number_moved